requirements.txt
matplotlib
numpy
pandas
pyyaml
seaborn
streamlit
requests
openpyxl
bokeh